Move avatar into user package
