+++ server/src/sessions/database.go
@@ -4,6 +4,8 @@ import (
 	"context"
 	"errors"
 
+	"scrumlr.io/server/identifiers"
+
 	"github.com/google/uuid"
 	"github.com/uptrace/bun"
 	"scrumlr.io/server/common"
