+++ server/src/users/avatar/accessories_type.go
