+++ server/src/users/avatar/clothe_type.go
