+++ server/src/users/avatar/eye_type.go
