+++ server/src/users/avatar/eyebrow_type.go
