+++ server/src/users/avatar/facial_hair_color.go
