+++ server/src/users/avatar/facial_hair_type.go
