+++ server/src/users/avatar/graphic_type.go
