+++ server/src/users/avatar/hair_color.go
