+++ server/src/users/avatar/skin_color.go
