+++ server/src/users/avatar/top_type.go
