+++ server/src/users/dto.go
@@ -16,16 +16,16 @@ type User struct {
 	Name string `json:"name"`
 
 	// The user's avatar configuration
-	Avatar *types.Avatar `json:"avatar,omitempty"`
+	Avatar *Avatar `json:"avatar,omitempty"`
 
 	// The user's account type configuration
 	AccountType types.AccountType `json:"accountType"`
 }
 
 type UserUpdateRequest struct {
-	ID     uuid.UUID     `json:"-"`
-	Name   string        `json:"name"`
-	Avatar *types.Avatar `json:"avatar,omitempty"`
+	ID     uuid.UUID `json:"-"`
+	Name   string    `json:"name"`
+	Avatar *Avatar   `json:"avatar,omitempty"`
 }
 
 func (u *User) From(user DatabaseUser) *User {
